+++ requirements-dev.txt
@@ -1,4 +1,9 @@
+-r requirements.txt
+
 flake8
-pytest
-coverage
+pytest >= 2.9.1
+coverage < 5.0
 pytest-cov
+setuptools >= 30.3.0
+setuptools_scm
+twine
