+++ requirements.txt
@@ -1,11 +1,1 @@
-enum34 ; python_version<='3.4'
-netaddr
-netifaces
-scapy>=2.4.3
-requests
-PrettyTable
-urllib3>=1.24.2,<1.25
-ruamel.yaml
-requests_mock
-future
-packaging
+-e .
