+++ setup.cfg
@@ -1,3 +1,56 @@
+[metadata]
+name = kube-hunter
+description = Kubernetes security weaknesses hunter for humans
+long_description = file: README.md
+long_description_content_type = text/markdown
+author = Aqua Security
+author_email = support@aquasec.com
+url = https://github.com/aquasecurity/kube-hunter
+keywords =
+    aquasec
+    hunter
+    kubernetes
+    k8s
+    security
+license_file = LICENSE
+classifiers =
+    Development Status :: 4 - Beta
+    Environment :: Console
+    License :: OSI Approved :: Apache Software License
+    Natural Language :: English
+    Operating System :: OS Independent
+    Programming Language :: Python :: 3.6
+    Programming Language :: Python :: 3.7
+    Programming Language :: Python :: 3.8
+    Topic :: Security
+
+[options]
+zip_safe = False
+packages = find:
+install_requires =
+    netaddr
+    netifaces
+    scapy>=2.4.3
+    requests
+    PrettyTable
+    urllib3>=1.24.2,<1.25
+    ruamel.yaml
+    requests_mock
+    future
+    packaging
+setup_requires =
+    setuptools>=30.3.0
+    setuptools_scm
+test_requires =
+    pytest>=2.9.1
+    coverage<5.0
+    pytest-cov
+python_requires = >=3.6
+
+[options.entry_points]
+console_scripts =
+    kube-hunter = kube_hunter.__main__:main
+
 [aliases]
 test=pytest
 
@@ -33,3 +86,7 @@ exclude_lines =
     if __name__ == .__main__.:
     # Don't complain about log messages not being tested
     logging\.
+
+# Files to exclude from consideration
+omit =
+    kube_hunter/__main__.py
