+++ setup.py
@@ -1,0 +1,5 @@
+from setuptools import setup
+
+setup(
+    use_scm_version=True,
+)
